# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * نظام تدقيق الحوالات API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransferStatus = typeof TransferStatus[keyof typeof TransferStatus];


export const TransferStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type TransferRiskLevel = typeof TransferRiskLevel[keyof typeof TransferRiskLevel];


export const TransferRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Transfer {
  id: number;
  operationNumber: string;
  amount: number;
  fromAccount: string;
  toAccount: string;
  recipientName: string;
  /** @nullable */
  comment?: string | null;
  status: TransferStatus;
  riskLevel: TransferRiskLevel;
  riskScore: number;
  agentId: number;
  agentName: string;
  /** @nullable */
  rejectionReason?: string | null;
  /** @nullable */
  transferDate?: string | null;
  createdAt: string;
}

export interface TransferInput {
  operationNumber: string;
  amount: number;
  fromAccount: string;
  toAccount: string;
  recipientName: string;
  comment?: string;
  agentId: number;
  riskScore: number;
  transferDate?: string;
}

export interface RejectInput {
  reason?: string;
}

export interface TransferStats {
  total: number;
  pending: number;
  approved: number;
  rejected: number;
  totalAmount: number;
  pendingAmount: number;
  approvedAmount: number;
  rejectedAmount: number;
}

export interface Agent {
  id: number;
  name: string;
  phone: string;
  balance: number;
  lastActivityAt: string;
  totalTransfers: number;
  pendingTransfers: number;
}

export interface InactiveAgent {
  id: number;
  name: string;
  phone: string;
  lastActivityAt: string;
  inactiveHours: number;
}

export interface AgentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
}

export interface ChangeAgentInput {
  agentId: number;
}

export interface StatementSummary {
  totalCount: number;
  approvedCount: number;
  pendingCount: number;
  rejectedCount: number;
  totalAmount: number;
  approvedAmount: number;
  pendingAmount: number;
  rejectedAmount: number;
}

export interface AgentStatement {
  agent: Agent;
  balance: number;
  summary: StatementSummary;
  transfers: Transfer[];
}

export interface ScanInput {
  imageBase64: string;
}

export interface ScanResult {
  /** @nullable */
  operationNumber: string | null;
  /** @nullable */
  amount: number | null;
  /** @nullable */
  fromAccount: string | null;
  /** @nullable */
  toAccount: string | null;
  /** @nullable */
  recipientName: string | null;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  transferDate: string | null;
  riskScore: number;
  confidence: number;
}

export interface Conversation {
  agentId: number;
  agentName: string;
  lastMessage: string;
  lastMessageAt: string;
  unreadCount: number;
}

export type MessageType = typeof MessageType[keyof typeof MessageType];


export const MessageType = {
  text: 'text',
  image: 'image',
} as const;

export type MessageDirection = typeof MessageDirection[keyof typeof MessageDirection];


export const MessageDirection = {
  incoming: 'incoming',
  outgoing: 'outgoing',
} as const;

export interface Message {
  id: number;
  agentId: number;
  content: string;
  type: MessageType;
  direction: MessageDirection;
  /** @nullable */
  imageUrl?: string | null;
  sentAt: string;
}

export type ListTransfersParams = {
status?: ListTransfersStatus;
/**
 * @nullable
 */
agentId?: number | null;
/**
 * @nullable
 */
search?: string | null;
};

export type ListTransfersStatus = typeof ListTransfersStatus[keyof typeof ListTransfersStatus];


export const ListTransfersStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;