# lib/api-zod/src/generated/types/agentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * نظام تدقيق الحوالات API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
}