# lib/api-zod/src/generated/types/agentStatement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * نظام تدقيق الحوالات API
 * OpenAPI spec version: 0.1.0
 */
import type { Agent } from './agent';
import type { StatementSummary } from './statementSummary';
import type { Transfer } from './transfer';

export interface AgentStatement {
  agent: Agent;
  balance: number;
  summary: StatementSummary;
  transfers: Transfer[];
}