# lib/api-zod/src/generated/types/changeAgentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * نظام تدقيق الحوالات API
 * OpenAPI spec version: 0.1.0
 */

export interface ChangeAgentInput {
  agentId: number;
}