# lib/api-zod/src/generated/types/statementSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * نظام تدقيق الحوالات API
 * OpenAPI spec version: 0.1.0
 */

export interface StatementSummary {
  totalCount: number;
  approvedCount: number;
  pendingCount: number;
  rejectedCount: number;
  totalAmount: number;
  approvedAmount: number;
  pendingAmount: number;
  rejectedAmount: number;
}